+++ npm-packages/cli/source/components/commands/preview.tsx
@@ -24,9 +24,18 @@ export default function Preview({cli}: {cli: Result<Flags>}) {
 				}
 			}
 
-			await execa('mkdir "node_modules/.pulse/server"', {
-				shell: true,
-			});
+			// Create node_modules/.pulse/server directory
+			if (!fs.existsSync('node_modules/.pulse/server')) {
+				if (process.platform === 'win32') {
+					await execa('mkdir node_modules\\.pulse\\server', {
+						shell: true,
+					});
+				} else {
+					await execa('mkdir -p node_modules/.pulse/server', {
+						shell: true,
+					});
+				}
+			}
 
 			if (process.platform === 'win32') {
 				await execa(