+++ web/app/(main-layout)/layout.tsx
@@ -25,21 +25,23 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en" suppressHydrationWarning>
-      <body className={`h-[100dvh] w-[100dvw] antialiased`}>
+      <body
+        className={`bg-default h-[100dvh] w-[100dvw] pt-[env(safe-area-inset-top)] pb-[env(safe-area-inset-bottom)] antialiased`}
+      >
         <Analytics />
         <Suspense>
           <WrappedHeroUIProvider>
             <EditorContextProvider>
               <CapacitorProvider>
                 <InterModuleCommunicationProvider>
-                    <DndProvider>
-                      <RemoteModuleProvider isPreventingCSS={true}>
-                        <InputDeviceProvider>
-                          <Toaster />
-                          <Nav>{children}</Nav>
-                        </InputDeviceProvider>
-                      </RemoteModuleProvider>
-                    </DndProvider>
+                  <DndProvider>
+                    <RemoteModuleProvider isPreventingCSS={true}>
+                      <InputDeviceProvider>
+                        <Toaster />
+                        <Nav>{children}</Nav>
+                      </InputDeviceProvider>
+                    </RemoteModuleProvider>
+                  </DndProvider>
                 </InterModuleCommunicationProvider>
               </CapacitorProvider>
             </EditorContextProvider>