+++ npm-packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @pulse-editor/shared-utils
 
+## 0.1.1-beta.64
+
+### Patch Changes
+
+- Update types
+
+## 0.1.1-beta.63
+
+### Patch Changes
+
+- Update packages
+
+## 0.1.1-beta.62
+
+### Patch Changes
+
+- Update to react 19.2.1
+
+## 0.1.1-beta.61
+
+### Patch Changes
+
+- Update packages and fix react versions
+
 ## 0.1.1-beta.60
 
 ### Patch Changes